update POST POI - add missing fields error exception

+++ routes/pois.py
@@ -44,7 +44,7 @@ def returnAllPois():
                 compteur += 1
             dictionnaire[cle] = valeur
     malistFormatBon.append(dictionnaire)
-    return jsonify({'pois': malistFormatBon})
+    return jsonify({'pois': malistFormatBon}), 200
 
 
 @routes.route('/api/pois/<int:idp>', methods=['GET'])
@@ -85,7 +85,7 @@ def returnOnepoi(idp):
                 compteur += 1
             dictionnaire[cle] = valeur
     malistFormatBon.append(dictionnaire)
-    return jsonify({'poi': malistFormatBon})
+    return jsonify({'poi': malistFormatBon}), 200
 
 
 @routes.route('/api/pois', methods=['POST'])
@@ -100,7 +100,7 @@ def addOnePoi():
     rjson = request.get_json()
     tour_id = rjson.get("tour_id", "null")
     typespois_id = rjson.get("typespois_id", "null")
-    
+
     if (tour_id != "null") :
         optionalVal['tour_id'] = request.json['tour_id']
     else:
@@ -134,48 +134,80 @@ def addOnePoi():
 
 
 
-@routes.route('/api/pois', methods=['PATCH'])
-def modifyOnePoiFieldValue():
+@routes.route('/api/pois/<int:idp>', methods=['PATCH'])
+def modifyOnePoiFieldValue(idp):
+    try:
+        currentPoi = models.Pois.query.filter_by(id = idp).first()
+        if(currentPoi == None):
+            raise ValueError('This POI does not exist')
+    except ValueError:
+        resp = jsonify({"error": 'This POI does not exist'})
+        resp.status_code = 404
+        return resp
 
-    id_value = request.json['id']
-    currentPoi = models.Pois.query.filter_by(id = id_value).first()
+    if(request.json is None):
+        return returnOnepoi(idp)
 
-    for key, value in request.json.items():
-        if key not in ['id']:
+    try:
+        contribExist = 0
+        for key, value in request.json.items():
             currentField = models.Fields.query.filter_by(name=key).first()
-            currentValue = models.Values(value=value)
-            currentContrib = models.Contributions(1, 'in progress',
+
+            allContribs = models.Contributions.query.all()
+            for contrib in allContribs:
+                if(contrib.idpoi == currentPoi.id and contrib.idfield == currentField.id):
+                    contribExist = 1;
+                    lastContrib = contrib;
+
+            if(contribExist):
+                currentValue = models.Values(value=value)
+                currentContrib = models.Contributions(lastContrib.version+1, 'in progress',
                                                currentPoi, currentField, currentValue)
-            db.session.add(currentContrib)
+                db.session.add(currentContrib)
+            else:
+                currentValue = models.Values(value=value)
+                currentContrib = models.Contributions(1, 'in progress',
+                                           currentPoi, currentField, currentValue)
+                db.session.add(currentContrib)
+        db.session.commit()
 
-    db.session.commit()
-    return "Modif. OK", 204
+        return jsonify({'poi_id': currentPoi.id, 'field_id': currentField.id, 'field_name': currentField.name,\
+                        'value_id': currentContrib.idvalue, 'value': value,  }), 200
+    except:
+        resp = jsonify({"error": 'Missing required fields or Bad key of entered name fields'})
+        resp.status_code = 403
+        return resp
 
-# lancer requete patch : http  PATCH http://localhost:5000/api/pois id=110 desc=newValue
+
+# lancer requete patch : http PATCH http://localhost:5000/api/pois/110 desc=newValue
 
 
 
 
 @routes.route('/api/pois/<int:idp>', methods=['DELETE'])
 def deleteOnePoi(idp):
+
     selectedPoi = models.Pois.query.filter_by(id=idp).first()
-    selectedContribs = models.Contributions.query.filter_by(idpoi=idp).all()
+    if(selectedPoi is not None):
+        selectedContribs = models.Contributions.query.filter_by(idpoi=idp).all()
 
-    for contrib in selectedContribs:
-        selectedPoisValues = models.Values.query.filter_by(id=contrib.idvalue).first()
-        db.session.delete(selectedPoisValues)
+        for contrib in selectedContribs:
+            selectedPoisValues = models.Values.query.filter_by(id=contrib.idvalue).first()
+            db.session.delete(selectedPoisValues)
 
-    db.session.commit()
-
-
-    db.session.delete(selectedPoi)
-    db.session.commit()
-
-    for contribution in selectedContribs:
-        db.session.delete(contribution)
-    db.session.commit()
+        db.session.commit()
 
 
-    return "SUPPR. OK", 200
+        db.session.delete(selectedPoi)
+        db.session.commit()
+
+        for contribution in selectedContribs:
+            db.session.delete(contribution)
+        db.session.commit()
+
+
+        return "SUPPR. OK", 204
+    else:
+        return jsonify({"error": 'This POI does not exists'}), 404
 
 # lancer requete delete : http  DELETE http://localhost:5000/api/pois/110
